API Reference

Pybtex.BibEntry — Type

This object represents a BibTeX entry. It is a wrapper around pybtex's pybtex.database.Entry object, connecting it with its corresponding key.

Pybtex.BibLibrary — Type

This provides a simple interface to read and manipulate bibtex databases.

Pybtex.BibType — Type

This module provides the types used in the BibLibrary module. These are the general BibTeX types useful for dispatching entries on them.

Pybtex.PersonName — Type

Module for handling person names.

Base.eltype — Method

This function returns the type of the BibTeX entry.

Base.getindex — Method

Returns the entry with the given key.

Base.getproperty — Method

Overload Base's getproperty function to allow for accessing the entries of the database and useful aliases.

Base.haskey — Method

Check if a key exists in the database.

Base.insert! — Method

Insert an entry into the database.

Base.iterate — Function

Iterate over the entries in the BibLibrary.

Base.keys — Method

Returns the keys of the entries in the database.

Base.length — Method

Returns the number of entries in the database.

Base.pop! — Method

Remove an entry from the database based on its key.

Base.show — Method

This function provides a simple way to print a BibEntry to the console.

Base.show — Method

This function provides a simple way to print a person's name to the console.

Base.values — Method

Return a vector of all values in the database

Pybtex.getADSURL — Method

This function returns the ADS URL of the publication.

Pybtex.getAbstract — Method

This function returns the abstract of the publication, if available.

Pybtex.getAllFields — Method

This function returns all fields of a BibTeX entry.

Pybtex.getArXiv — Method

Get arXiv number, assuming it is in the eprint field.

Pybtex.getAuthors — Method

This function returns the authors of a BibTeX entry. The authors are returned as a vector of PersonName objects.

If the authors field is not present, but editors are found, the editors are returned instead.

Pybtex.getBibTeX — Method

This function returns the BibTeX representation of the entry.

Pybtex.getDOI — Method

This function returns the DOI of the publication.

Pybtex.getEId — Method

This function returns the electronic identifier of the publication.

Pybtex.getEdition — Method

This function returns the edition of the publication.

Pybtex.getEditors — Method

This function returns the editors of a BibTeX entry. The editors are returned as a vector of PersonName objects.

Pybtex.getEntry — Method

This function returns a BibTeX entry from a BibTeX library.

Pybtex.getFileName — Method

This function returns the file name(s) of the publication.

Pybtex.getISBN — Method

This function returns the ISBN of the publication.

Pybtex.getJournal — Method

This function returns the journal where the publication was published.

Pybtex.getKeywords — Method

This function returns the keywords of the publication, if available.

Pybtex.getLocation — Method

This function returns the location of the publication.

Pybtex.getNumber — Method

This function returns the number of the publication.

Pybtex.getPages — Method

This function returns the pages of the publication.

Pybtex.getPublisher — Method

This function returns the publisher of the publication.

Pybtex.getSchool — Method

This function returns the school of the publication, if available. This is common for theses.

Pybtex.getType — Method

This function returns the type of a BibTeX entry.

Pybtex.getTypeName — Method

Returns the name of the type of the entry as a string.

Pybtex.getURL — Method

This function returns the URL of the publication.

Pybtex.getVolume — Method

This function returns the volume of the publication.

Pybtex.getYear — Method

This function returns the year of the publication.

Pybtex.hasField — Method

This function checks if a field is present in a BibTeX entry.

Pybtex.journalAbbreviationFromTag — Method

Returns the journal abbreviation from the tag. For example, the tag "apj" returns "Astrophys. J.".

Pybtex.journalFullNameFromTag — Method

Returns the journal full name from the tag. For example, the tag "apj" returns "Astrophysical Journal".

Pybtex.journalTagFromFullName — Method

Returns the journal tag from the full name. For example, the full name "Astrophysical Journal" returns "apj". This is essentially an inversion of journalFullNameFromTag.

Pybtex.numberOfAuthors — Method

This function returns the number of authors of a BibTeX entry.

Pybtex.numberOfEditors — Method

This function returns the number of editors of a BibTeX entry.

Pybtex.pybtexToPersonName — Method

Converts a Pybtex person object to a PersonName object.

Pybtex.readBibtexDataBase — Method

Reads a bibtex database from a file and returns a BibLibrary object.

Pybtex.removeCurlyBracesLimiters — Method

This function fixes the strings from the bibtex file.

Pybtex.stringPy2Jl — Method

This unexported function converts a python string to a julia string.